Implemented profile picture uploading.
I was running into an issue when I verified the token used in the POST request. The hashed value was always different from the value in the database, even when the account I was using only had one token. I resolved this by accessing the token within the php code, and this resolved the error.

I think that sending the random bytes through a post request somehow changed it, making the hashed value different.