Fixed usernames expanding the result infinitely
Previously, the results would get as wide as they needed to be to fit the entire long usernme, but now, there is a max width, and any ovrflow is indicated by an ellipsis.

I also removed the visible hyperlink for result. The text is now black even after you click it.