Fixed a bug with notebook search
When there was no notebooks matching the search, the backend wasn't responding with "success", so the web page didn't know to update the results list.
This meant that if you searched for a notebook that exists, then you searched for one that didn't, the result would still show the results from the previous search.